Update README.md
Open the index.html file in your web browser. You'll see the main page with a navigation bar.

Click the links in the navigation bar to load different web pages into the iframe. Notice that each page is loaded and displayed within the iframe while maintaining isolation.

Check the browser's developer console and network tab to observe the caching mechanism at work. Pages are fetched and cached the first time you visit them.

Customization
You can easily customize this project by adding more menu items to the navigation bar and specifying the URLs for the pages you want to display in the iframes. Just edit the HTML and JavaScript in the index.html file.